Migration of GetInfo to cpp files for statements
