+++ src/include/parser/copy_statement.h
@@ -36,28 +36,9 @@ class CopyStatement : public SQLStatement {
 
   virtual void Accept(SqlNodeVisitor *v) override { v->Visit(this); }
 
-  const std::string GetInfo(int num_indent) const override {
-    std::ostringstream os;
-    os << StringUtil::Indent(num_indent) << "CopyStatement\n";
-    os << StringUtil::Indent(num_indent + 1)
-       << "-> Type :: " << CopyTypeToString(type) << "\n";
-    os << cpy_table.get()->GetInfo(num_indent + 1) << std::endl;
+  const std::string GetInfo(int num_indent) const override;
 
-    os << StringUtil::Indent(num_indent + 1) << "-> File Path :: " << file_path
-       << std::endl;
-    os << StringUtil::Indent(num_indent + 1) << "-> Delimiter :: " << delimiter;
-    return os.str();
-  }
-
-  const std::string GetInfo() const override {
-    std::ostringstream os;
-
-    os << "SQLStatement[COPY]\n";
-
-    os << GetInfo(1);
-
-    return os.str();
-  }
+  const std::string GetInfo() const override;
 
   std::unique_ptr<TableRef> cpy_table;
