+++ src/include/parser/create_statement.h
@@ -225,116 +225,9 @@ class CreateStatement : public TableRefStatement {
 
   virtual void Accept(SqlNodeVisitor *v) override { v->Visit(this); }
 
-  const std::string GetInfo(int num_indent) const override {
-    std::ostringstream os;
-    os << StringUtil::Indent(num_indent) << "CreateStatement\n";
-    os << StringUtil::Indent(num_indent + 1);
+  const std::string GetInfo(int num_indent) const override;
 
-    switch (type) {
-      case CreateStatement::CreateType::kTable: {
-        os << "Create type: Table"
-           << "\n";
-        os << StringUtil::Indent(num_indent + 1)
-           << StringUtil::Format("IF NOT EXISTS: %s",
-                                 (if_not_exists) ? "True" : "False") << "\n";
-        os << StringUtil::Indent(num_indent + 1)
-           << StringUtil::Format("Table name: %s", GetTableName().c_str())
-           << "\n";
-        break;
-      }
-      case CreateStatement::CreateType::kDatabase: {
-        os << "Create type: Database"
-           << "\n";
-        os << StringUtil::Indent(num_indent + 1)
-           << StringUtil::Format("Database name: %s", GetDatabaseName().c_str())
-           << "\n";
-        break;
-      }
-      case CreateStatement::CreateType::kIndex: {
-        os << "Create type: Index"
-           << "\n";
-        os << StringUtil::Indent(num_indent + 1) << index_name << "\n";
-        os << StringUtil::Indent(num_indent + 1)
-           << "INDEX : table : " << GetTableName() << " unique : " << unique
-           << " attrs : ";
-        for (auto &key : index_attrs) os << key << " ";
-        os << "\n";
-        os << StringUtil::Indent(num_indent + 1)
-           << "Type : " << IndexTypeToString(index_type) << "\n";
-        os << "\n";
-        break;
-      }
-      case CreateStatement::CreateType::kTrigger: {
-        os << "Create type: Trigger"
-           << "\n";
-        os << StringUtil::Indent(num_indent + 1)
-           << StringUtil::Format("Trigger table name: %s", trigger_name.c_str())
-           << "\n";
-        os << StringUtil::Indent(num_indent + 1)
-           << StringUtil::Format("Trigger name: %s", GetTableName().c_str())
-           << "\n";
-        break;
-      }
-      case CreateStatement::CreateType::kSchema: {
-        os << "Create type: Schema"
-           << "\n";
-        os << StringUtil::Indent(num_indent + 1)
-           << StringUtil::Format("Schema name: %s", schema_name.c_str())
-           << "\n";
-        break;
-      }
-      case CreateStatement::CreateType::kView: {
-        os << "Create type: View"
-           << "\n";
-        os << StringUtil::Indent(num_indent + 1)
-           << StringUtil::Format("View name: %s", view_name.c_str()) << "\n";
-        break;
-      }
-    }
-
-    if (!columns.empty()) {
-      for (auto &col : columns) {
-        if (col->name.empty()) {
-          continue;
-        }
-        if (col->type == ColumnDefinition::DataType::PRIMARY) {
-          os << StringUtil::Indent(num_indent + 1) << "-> PRIMARY KEY : ";
-          for (auto &key : col->primary_key) os << key << " ";
-          os << "\n";
-        } else if (col->type == ColumnDefinition::DataType::FOREIGN) {
-          os << StringUtil::Indent(num_indent + 1)
-             << "-> FOREIGN KEY : References " << col->name << " Source : ";
-          for (auto &key : col->foreign_key_source) {
-            os << key << " ";
-          }
-          os << "Sink : ";
-          for (auto &key : col->foreign_key_sink) {
-            os << key << " ";
-          }
-          os << "\n";
-        } else {
-          os << StringUtil::Indent(num_indent + 1)
-             << "-> COLUMN REF : " << col->name << " "
-             // << col->type << " not null : "
-             << col->not_null << " primary : " << col->primary << " unique "
-             << col->unique << " varlen " << col->varlen << "\n";
-        }
-      }
-    }
-    std::string info = os.str();
-    StringUtil::RTrim(info);
-    return info;
-  }
-
-  const std::string GetInfo() const override {
-    std::ostringstream os;
-
-    os << "SQLStatement[CREATE]\n";
-
-    os << GetInfo(1);
-
-    return os.str();
-  }
+  const std::string GetInfo() const override;
 
   CreateType type;
   bool if_not_exists;
