+++ src/include/parser/transaction_statement.h
@@ -35,33 +35,9 @@ class TransactionStatement : public SQLStatement {
 
   virtual void Accept(SqlNodeVisitor *v) override { v->Visit(this); }
 
-  const std::string GetInfo(int num_indent) const override {
-    std::ostringstream os;
-    os << StringUtil::Indent(num_indent) << "TransactionStatement\n";
-    os << StringUtil::Indent(num_indent + 1) << "Type: ";
+  const std::string GetInfo(int num_indent) const override;
 
-    switch (type) {
-      case kBegin:
-        os << "Begin\n";
-        break;
-      case kCommit:
-        os << "Commit\n";
-        break;
-      case kRollback:
-        os << "Rollback\n";
-        break;
-    }
-
-    return os.str();
-  }
-
-  const std::string GetInfo() const override {
-    std::ostringstream os;
-    os << "SQLStatement[TRANSACTION]\n";
-    os << GetInfo(1);
-
-    return os.str();
-  }
+  const std::string GetInfo() const override;
 
   CommandType type;
 };
